+++ src/problem_6.rs
@@ -1,7 +1,7 @@
 //! Sum square difference
 
-pub fn solve() -> i32 {
-    (sum_of_natural_numbers(100).pow(2) - sum_of_squared_natural_numbers(100)) as i32
+pub fn solve() -> u64 {
+    (sum_of_natural_numbers(100).pow(2) - sum_of_squared_natural_numbers(100)) as u64
 }
 
 pub fn sum_of_natural_numbers(n: u32) -> u32 {
